+++ javascript/node/selenium-webdriver/builder.js
@@ -113,6 +113,9 @@ var Builder = function() {
   /** @private {ie.Options} */
   this.ieOptions_ = null;
 
+  /** @private {edge.Options} */
+  this.edgeOptions_ = null;
+
   /** @private {safari.Options} */
   this.safariOptions_ = null;
 
@@ -338,7 +341,7 @@ Builder.prototype.setOperaOptions = function(options) {
 
 
 /**
- * Sets Internet Explorer specific
+ * Sets Microsoft's Internet Explorer specific
  * {@linkplain selenium-webdriver/ie.Options options} for drivers created by
  * this builder. Any proxy settings defined on the given options will take
  * precedence over those set through {@link #setProxy}.
@@ -351,6 +354,19 @@ Builder.prototype.setIeOptions = function(options) {
   return this;
 };
 
+/**
+ * Sets Microsoft's Edge specific
+ * {@linkplain selenium-webdriver/edge.Options options} for drivers created by
+ * this builder. Any proxy settings defined on the given options will take
+ * precedence over those set through {@link #setProxy}.
+ *
+ * @param {!edge.Options} options The MicrosoftEdgeDriver options to use.
+ * @return {!Builder} A self reference.
+ */
+Builder.prototype.setEdgeOptions = function(options) {
+  this.edgeOptions_ = options;
+  return this;
+};
 
 /**
  * Sets Safari specific {@linkplain selenium-webdriver/safari.Options options}
@@ -428,6 +444,9 @@ Builder.prototype.build = function() {
 
   } else if (browser === Browser.SAFARI && this.safariOptions_) {
     capabilities.merge(this.safariOptions_.toCapabilities());
+
+  } else if (browser === Browser.EDGE && this.edgeOptions_) {
+    capabilities.merge(this.edgeOptions_.toCapabilities());
   }
 
   // Check for a remote browser.
@@ -464,6 +483,12 @@ Builder.prototype.build = function() {
       // index -> builder -> ie -> index
       var ie = require('./ie');
       return new ie.Driver(capabilities, this.flow_);
+
+    case Browser.EDGE:
+      // Requiring 'edge' above would create a cycle:
+      // index -> builder -> edge -> index
+      var edge = require('./edge');
+      return new edge.Driver(capabilities, null, this.flow_);
 
     case Browser.OPERA:
       // Requiring 'opera' would create a cycle: